[#6] Error: 작품 값 넘겨지지 않음
